Reducing the duration of a permanent session to 1 hour
Reducing the session duration allows for custom temporary configurations to
be reset within a reasonable time period. For instance, the values of
profile and date range filters persist throughout the length of the
session. If the session never expires, this data will still be loaded
the next time the user accesses the app, even if they log in after several days.
Resetting these runtime configurations, especially the date ranges,
allows the user to see the newest data when they log back in after a
long period of inactivity.